Progress bar func consolidate to one file
