+++ progress.go
@@ -25,6 +25,48 @@ type ProgressBar interface {
 	DrawColoredString(s string, color int) string
 }
 
+type ProgressBarConfigOptions func(*ProgressBarConfig)
+
+func WithProgressBarWidth(width int) ProgressBarConfigOptions {
+	return func(p *ProgressBarConfig) {
+		p.Width = width
+	}
+}
+
+func WithProgressFillCharacter(s string) ProgressBarConfigOptions {
+	return func(p *ProgressBarConfig) {
+		p.FillCharacter = s
+	}
+}
+
+func WithProgressRemaingCharacter(s string) ProgressBarConfigOptions {
+	return func(p *ProgressBarConfig) {
+		p.RemainingCharacter = s
+	}
+}
+
+func (f *ProgressBarConfig) DrawColoredString(s string, color int) string {
+	coloredString := fmt.Sprintf("\x1b[%dm%s\x1b[0m", color, s)
+	return coloredString
+}
+
+func NewProgressBarConfig(opts ...ProgressBarConfigOptions) *ProgressBarConfig {
+	const (
+		width = 50
+	)
+	progBarConf := &ProgressBarConfig{Width: width}
+	progBarConf.FillCharacter = progBarConf.DrawColoredString("#", 92)
+	progBarConf.RemainingCharacter = progBarConf.DrawColoredString("-", 96)
+
+	for _, opt := range opts {
+
+		opt(progBarConf)
+
+	}
+
+	return progBarConf
+}
+
 func (f *FileCopyJob) DrawProgressBar() {
 	if f.ProgressCompleted < 0 {
 		f.ProgressCompleted = 0
